# api/openapi/spec.yaml
openapi: "3.0.3"
info:
  title: Paily Connect Core
  version: "1.0.0"
  description: >
    Proxy node aggregation, filtering and distribution service.
    Fetcher and Checker authenticate with Bearer service-secret.
    Admin endpoints accept either a Bearer JWT (obtained via /api/v1/auth/login)
    or the shared service-secret Bearer token.
servers:
  - url: /
  # Default security: admin endpoints accept AdminJWT OR ServiceBearer (OR = separate list entries).
  # Public endpoints and ServiceBearer-only endpoints override this with their own security field.
security:
  - AdminJWT: []
  - ServiceBearer: []
components:
  securitySchemes:
    AdminJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ServiceBearer:
      type: http
      scheme: bearer
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
paths:
...
  /{format}:
    get:
      operationId: Distribute
      tags: [distribution]
      summary: Get subscription file in the requested format
      parameters:
        - {
            name: format,
            in: path,
            required: true,
            schema: { type: string, enum: [clash, singbox, base64] },
          }
        - {
            name: preset,
            in: query,
            schema:
              { type: string, enum: [common, all, other], default: common },
            description: "Node selection preset: common (HK/TW/SG/US/JP priority), all (all regions, UN last), other (non-common regions only)",
          }
        - {
            name: filter_unlock,
            in: query,
            schema: { type: string },
            description: "Vertical bar-separated streaming service names",
          }
        - {
            name: filter_mode,
            in: query,
            schema: { type: string, enum: [select, mark], default: select },
          }
        - {
            name: tag,
            in: query,
            schema: { type: string },
            description: "Filter by checker tag",
          }
      responses:
        "200":
          description: Subscription file content
          content:
            text/plain:
              schema: { type: string }
            text/yaml:
              schema: { type: string }
            application/json:
              schema: { type: object }
        "400":
          description: Unknown format
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: No alive nodes
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }